API 리스트 UI 개선 #44
